CR review follow-up
