+++ src/main/java/hudson/plugins/gradle/BuildScanPublisher.java
@@ -2,8 +2,8 @@ package hudson.plugins.gradle;
 
 import hudson.Extension;
 import hudson.model.Run;
-import hudson.plugins.gradle.config.GlobalConfig;
-import hudson.util.Secret;
+import hudson.plugins.gradle.enriched.EnrichedSummaryConfig;
+import hudson.plugins.gradle.enriched.ScanDetailService;
 import org.jenkinsci.plugins.workflow.steps.Step;
 import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepDescriptor;
@@ -13,7 +13,6 @@ import org.kohsuke.stapler.DataBoundConstructor;
 
 import javax.annotation.Nonnull;
 import java.io.BufferedReader;
-import java.net.URI;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -39,9 +38,9 @@ public class BuildScanPublisher extends Step {
         @Override
         protected List<String> run() throws Exception {
             Run run = getContext().get(Run.class);
-            Secret buildScanAccessKey = GlobalConfig.get().getBuildScanAccessKey();
-            URI buildScanServer = GlobalConfig.get().getBuildScanServerUri();
-            BuildScanLogScanner scanner = new BuildScanLogScanner(new DefaultBuildScanPublishedListener(run, buildScanAccessKey, buildScanServer));
+            ScanDetailService scanDetailService = new ScanDetailService(EnrichedSummaryConfig.get());
+
+            BuildScanLogScanner scanner = new BuildScanLogScanner(new DefaultBuildScanPublishedListener(run, scanDetailService));
             try (
                     BufferedReader logReader = new BufferedReader(run.getLogReader());
                     Stream<String> lines = logReader.lines()
