+++ src/main/java/hudson/plugins/gradle/WithGradleExecution.java
@@ -2,9 +2,9 @@ package hudson.plugins.gradle;
 
 import hudson.model.Run;
 import hudson.model.TaskListener;
-import hudson.plugins.gradle.config.GlobalConfig;
+import hudson.plugins.gradle.enriched.EnrichedSummaryConfig;
 import hudson.plugins.gradle.enriched.ScanDetail;
-import hudson.plugins.gradle.enriched.ScanDetailServiceDefaultImpl;
+import hudson.plugins.gradle.enriched.ScanDetailService;
 import org.jenkinsci.plugins.workflow.graph.FlowNode;
 import org.jenkinsci.plugins.workflow.log.TaskListenerDecorator;
 import org.jenkinsci.plugins.workflow.steps.BodyExecutionCallback;
@@ -16,6 +16,7 @@ import java.io.PrintStream;
 import java.io.UncheckedIOException;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 public class WithGradleExecution extends StepExecution {
 
@@ -72,11 +73,11 @@ public class WithGradleExecution extends StepExecution {
                 BuildScanAction buildScanAction = existingAction == null
                         ? new BuildScanAction()
                         : existingAction;
-                ScanDetailServiceDefaultImpl scanDetailService = new ScanDetailServiceDefaultImpl(GlobalConfig.get().getBuildScanAccessKey(), GlobalConfig.get().getBuildScanServerUri());
+                ScanDetailService scanDetailService = new ScanDetailService(EnrichedSummaryConfig.get());
                 buildScans.forEach(scanUrl -> {
                     buildScanAction.addScanUrl(scanUrl);
-                    ScanDetail scanDetail = scanDetailService.getScanDetail(scanUrl);
-                    buildScanAction.addScanDetail(scanDetail);
+                    Optional<ScanDetail> scanDetail = scanDetailService.getScanDetail(scanUrl);
+                    scanDetail.ifPresent(buildScanAction::addScanDetail);
                 });
                 if (existingAction == null) {
                     run.addAction(buildScanAction);
