+++ src/main/java/hudson/plugins/gradle/enriched/HttpClientFactory.java
@@ -1,18 +1,15 @@
 package hudson.plugins.gradle.enriched;
 
-import hudson.plugins.gradle.config.GlobalConfig;
+import com.google.common.util.concurrent.Uninterruptibles;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 
-public class HttpClientProviderDefaultImpl implements HttpClientProvider {
+import java.util.concurrent.TimeUnit;
 
-    @Override
-    public CloseableHttpClient buildHttpClient() {
-        int httpClientTimeoutInSeconds = GlobalConfig.get().getHttpClientTimeoutInSeconds();
-        int httpClientMaxRetries = GlobalConfig.get().getHttpClientMaxRetries();
-        int httpClientDelayBetweenRetriesInSeconds = GlobalConfig.get().getHttpClientDelayBetweenRetriesInSeconds();
+class HttpClientFactory {
 
+    public CloseableHttpClient buildHttpClient(int httpClientTimeoutInSeconds, int httpClientMaxRetries, int httpClientDelayBetweenRetriesInSeconds) {
         RequestConfig config = RequestConfig.custom()
                 .setConnectTimeout(httpClientTimeoutInSeconds * 1000)
                 .setConnectionRequestTimeout(httpClientTimeoutInSeconds * 1000)
@@ -24,12 +21,7 @@ public class HttpClientProviderDefaultImpl implements HttpClientProvider {
                     if (executionCount > httpClientMaxRetries) {
                         return false;
                     } else {
-                        try {
-                            // Sleep before retrying
-                            Thread.sleep(httpClientDelayBetweenRetriesInSeconds * 1000L);
-                        } catch (InterruptedException ignore) {
-                        }
-
+                        Uninterruptibles.sleepUninterruptibly(httpClientDelayBetweenRetriesInSeconds, TimeUnit.SECONDS);
                         return true;
                     }
                 }).build();
