--- src/main/java/hudson/plugins/gradle/enriched/HttpClientProvider.java
@@ -1,9 +1,0 @@
-package hudson.plugins.gradle.enriched;
-
-import org.apache.http.impl.client.CloseableHttpClient;
-
-public interface HttpClientProvider {
-
-    CloseableHttpClient buildHttpClient();
-
-}
