+++ src/main/java/hudson/plugins/gradle/enriched/ScanDetail.java
@@ -1,25 +1,42 @@
 package hudson.plugins.gradle.enriched;
 
+import com.fasterxml.jackson.annotation.JsonAlias;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.kohsuke.stapler.export.ExportedBean;
 
+import java.util.List;
+
 @ExportedBean
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ScanDetail {
 
-    private String projectName;
-    private String buildToolType;
-    private String buildToolVersion;
-    private String requestedTasks;
-    private String hasFailed;
-    private String url;
+    public enum BuildToolType {
+        @JsonProperty("gradle")
+        GRADLE,
+        @JsonProperty("maven")
+        MAVEN;
+    }
 
-    private ScanDetail() {
+    private final String url;
+
+    @JsonAlias({"rootProjectName", "topLevelProjectName" })
+    private String projectName;
+    private BuildToolType buildToolType;
+    private String buildToolVersion;
+    @JsonAlias({"requestedTasks", "requestedGoals" })
+    private List<String> tasks;
+    private String hasFailed;
+
+    ScanDetail(String url) {
+        this.url = url;
     }
 
     public String getProjectName() {
         return projectName;
     }
 
-    public String getBuildToolType() {
+    public BuildToolType getBuildToolType() {
         return buildToolType;
     }
 
@@ -27,8 +44,8 @@ public class ScanDetail {
         return buildToolVersion;
     }
 
-    public String getRequestedTasks() {
-        return requestedTasks;
+    public List<String> getTasks() {
+        return tasks;
     }
 
     public boolean getHasFailed() {
@@ -39,56 +56,5 @@ public class ScanDetail {
         return url;
     }
 
-    static class ScanDetailBuilder {
-
-        private String projectName;
-        private String buildToolType;
-        private String buildToolVersion;
-        private String requestedTasks;
-        private String hasFailed;
-        private String url;
-
-        public ScanDetailBuilder withProjectName(String projectName) {
-            this.projectName = projectName;
-            return this;
-        }
-
-        public ScanDetailBuilder withBuildToolType(String buildToolType) {
-            this.buildToolType = buildToolType;
-            return this;
-        }
-
-        public ScanDetailBuilder withBuildToolVersion(String buildToolVersion) {
-            this.buildToolVersion = buildToolVersion;
-            return this;
-        }
-
-        public ScanDetailBuilder withRequestedTasks(String requestedTasks) {
-            this.requestedTasks = requestedTasks;
-            return this;
-        }
-
-        public ScanDetailBuilder withHasFailed(String hasFailed) {
-            this.hasFailed = hasFailed;
-            return this;
-        }
-
-        public ScanDetailBuilder withUrl(String url) {
-            this.url = url;
-            return this;
-        }
-
-        public ScanDetail build() {
-            ScanDetail scanDetail = new ScanDetail();
-            scanDetail.buildToolType = buildToolType;
-            scanDetail.buildToolVersion = buildToolVersion;
-            scanDetail.projectName = projectName;
-            scanDetail.requestedTasks = requestedTasks;
-            scanDetail.hasFailed = hasFailed;
-            scanDetail.url = url;
-            return scanDetail;
-        }
-
-    }
 }
