+++ src/main/java/hudson/plugins/gradle/injection/GradleBuildScanInjection.java
@@ -6,6 +6,7 @@ import com.google.common.base.Suppliers;
 import hudson.EnvVars;
 import hudson.FilePath;
 import hudson.model.Node;
+import hudson.plugins.gradle.config.GlobalConfig;
 import hudson.plugins.gradle.util.CollectionUtil;
 import hudson.remoting.VirtualChannel;
 
@@ -57,7 +58,7 @@ public class GradleBuildScanInjection implements BuildScanInjection {
 
     @Override
     public boolean isEnabled(Node node) {
-        InjectionConfig config = InjectionConfig.get();
+        GlobalConfig config = GlobalConfig.get();
 
         if (config.isDisabled() || isMissingRequiredParameters(config)) {
             return false;
