--- src/main/resources/hudson/plugins/gradle/config/GlobalConfig/help-buildScanAccessKey.html
@@ -1,3 +1,0 @@
-<div>
-    The access key for authenticating with the Gradle Enterprise server.
-</div>
