--- src/main/resources/hudson/plugins/gradle/config/GlobalConfig/help-enrichedSummaryEnabled.html
@@ -1,3 +1,0 @@
-<div>
-    Enable enriched build scan
-</div>
