--- src/main/resources/hudson/plugins/gradle/config/GlobalConfig/help-httpClientDelayBetweenRetriesInSeconds.html
@@ -1,3 +1,0 @@
-<div>
-    HTTP client delay between retries in seconds for requests to Gradle Enterprise API (default to 1).
-</div>
