--- src/main/resources/hudson/plugins/gradle/config/GlobalConfig/help-httpClientMaxRetries.html
@@ -1,3 +1,0 @@
-<div>
-    HTTP client max retries for requests to Gradle Enterprise API (default to 3).
-</div>
