--- src/main/resources/hudson/plugins/gradle/config/GlobalConfig/help-server.html
@@ -1,3 +1,0 @@
-<div>
-    The URL of the Gradle Enterprise server.
-</div>
