+++ src/main/resources/hudson/plugins/gradle/injection/InjectionConfig/config.jelly
@@ -1,32 +1,7 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form" xmlns:l="/lib/layout">
     <f:section title="${%Gradle Enterprise integration}">
-
-        <f:optionalBlock field="enrichedSummaryEnabled" title="${%Enable enriched summary}" inline="true">
-
-            <f:entry title="${%Gradle Enterprise server URL override}" field="buildScanServer">
-                <f:textbox />
-            </f:entry>
-
-            <f:entry title="${%Gradle Enterprise access key for build scan values}" field="buildScanAccessKey">
-                <f:password/>
-            </f:entry>
-
-            <f:section title="${%HTTP client settings}">
-                <f:entry title="${%Timeout in seconds}" field="httpClientTimeoutInSeconds">
-                    <f:textbox default="1"/>
-                </f:entry>
-                <f:entry title="${%Maximum retries}" field="httpClientMaxRetries">
-                    <f:textbox default="3"/>
-                </f:entry>
-                <f:entry title="${%Delay between retries in seconds}" field="httpClientDelayBetweenRetriesInSeconds">
-                    <f:textbox default="1"/>
-                </f:entry>
-            </f:section>
-
-        </f:optionalBlock>
-
-        <f:optionalBlock field="injectionEnabled" title="${%Enable auto-injection}" inline="true">
+        <f:optionalBlock field="enabled" title="${%Enable auto-injection}" inline="true">
 
             <j:if test="${instance.showLegacyConfigurationWarning}">
                 <f:entry>
