+++ src/main/resources/hudson/plugins/gradle/injection/InjectionConfig/help-accessKey.html
