+++ src/main/resources/hudson/plugins/gradle/injection/InjectionConfig/help-gradlePluginVersion.html
