+++ src/main/resources/hudson/plugins/gradle/injection/InjectionConfig/help-injectionEnabled.html
