+++ src/test/groovy/hudson/plugins/gradle/injection/BaseInjectionIntegrationTest.groovy
@@ -3,7 +3,6 @@ package hudson.plugins.gradle.injection
 import hudson.EnvVars
 import hudson.model.Label
 import hudson.plugins.gradle.AbstractIntegrationTest
-import hudson.plugins.gradle.config.GlobalConfig
 import hudson.slaves.DumbSlave
 import hudson.slaves.EnvironmentVariablesNodeProperty
 import hudson.slaves.NodeProperty
@@ -31,8 +30,8 @@ class BaseInjectionIntegrationTest extends AbstractIntegrationTest {
         env
     }
 
-    GlobalConfig withInjectionConfig(@DelegatesTo(GlobalConfig) Closure closure) {
-        def config = GlobalConfig.get()
+    InjectionConfig withInjectionConfig(@DelegatesTo(InjectionConfig) Closure closure) {
+        def config = InjectionConfig.get()
 
         closure.setDelegate(config)
         closure.run()
