+++ src/test/groovy/hudson/plugins/gradle/injection/BuildScanInjectionGradleIntegrationTest.groovy
@@ -47,7 +47,7 @@ class BuildScanInjectionGradleIntegrationTest extends BaseGradleInjectionIntegra
         def page = webClient.goTo("configure")
         def form = page.getFormByName("config")
 
-        form.getInputByName("_.injectionEnabled").click()
+        form.getInputByName("_.enabled").click()
         form.getInputByName("_.server").setValueAttribute("https://localhost")
         form.getInputByName("_.gradlePluginVersion").setValueAttribute("3.11.1")
 
@@ -430,7 +430,7 @@ class BuildScanInjectionGradleIntegrationTest extends BaseGradleInjectionIntegra
 
         when:
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             server = "http://localhost"
             gradlePluginVersion = GRADLE_ENTERPRISE_PLUGIN_VERSION
         }
@@ -452,7 +452,7 @@ class BuildScanInjectionGradleIntegrationTest extends BaseGradleInjectionIntegra
 
         when:
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             server = null
             gradlePluginVersion = null
         }
@@ -488,7 +488,7 @@ class BuildScanInjectionGradleIntegrationTest extends BaseGradleInjectionIntegra
 
         when:
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             server = 'http://localhost'
             allowUntrusted = true
             gradlePluginVersion = GRADLE_ENTERPRISE_PLUGIN_VERSION
@@ -513,7 +513,7 @@ class BuildScanInjectionGradleIntegrationTest extends BaseGradleInjectionIntegra
 
         when:
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             server = null
             allowUntrusted = false
             gradlePluginVersion = null
@@ -610,7 +610,7 @@ task hello {
         }
 
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             gradlePluginVersion = GRADLE_ENTERPRISE_PLUGIN_VERSION
             gradlePluginRepositoryUrl = repositoryAddress?.toString()
         }
@@ -620,7 +620,7 @@ task hello {
 
     private void disableBuildInjection(DumbSlave slave) {
         withInjectionConfig {
-            injectionEnabled = false
+            enabled = false
         }
 
         restartSlave(slave)
@@ -628,7 +628,7 @@ task hello {
 
     private void turnOffBuildInjection(DumbSlave slave) {
         withInjectionConfig {
-            injectionEnabled = true
+            enabled = true
             gradlePluginVersion = null
         }
