+++ src/test/groovy/hudson/plugins/gradle/injection/InjectionConfigTest.groovy
@@ -2,7 +2,6 @@ package hudson.plugins.gradle.injection
 
 import com.gargoylesoftware.htmlunit.html.HtmlButton
 import com.gargoylesoftware.htmlunit.html.HtmlForm
-import hudson.plugins.gradle.config.GlobalConfig
 import hudson.slaves.EnvironmentVariablesNodeProperty
 import hudson.util.FormValidation
 import hudson.util.XStream2
@@ -10,11 +9,11 @@ import spock.lang.Shared
 import spock.lang.Subject
 import spock.lang.Unroll
 
-@Subject(GlobalConfig.class)
+@Subject(InjectionConfig.class)
 class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
 
     @Shared
-    FilenameFilter injectionConfigXmlFilter = { _, name -> name == "hudson.plugins.gradle.config.GlobalConfig.xml" }
+    FilenameFilter injectionConfigXmlFilter = { _, name -> name == "hudson.plugins.gradle.injection.InjectionConfig.xml" }
 
     @Unroll
     def "sets showLegacyConfigurationWarning to true if any of legacy env variables is set"() {
@@ -29,7 +28,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         j.jenkins.getGlobalNodeProperties().add(env)
 
         expect:
-        GlobalConfig.get().isShowLegacyConfigurationWarning() == showWarning
+        InjectionConfig.get().isShowLegacyConfigurationWarning() == showWarning
 
         where:
         name                                                           | value               || showWarning
@@ -53,7 +52,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
     @Unroll
     def "validates server url"() {
         expect:
-        with(GlobalConfig.get().doCheckServer(url)) {
+        with(InjectionConfig.get().doCheckServer(url)) {
             kind == expectedKind
             message == expectedMssage
         }
@@ -74,7 +73,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
     @Unroll
     def "validates plugin repository url"() {
         expect:
-        with(GlobalConfig.get().doCheckGradlePluginRepositoryUrl(url)) {
+        with(InjectionConfig.get().doCheckGradlePluginRepositoryUrl(url)) {
             kind == expectedKind
             message == expectedMssage
         }
@@ -95,11 +94,11 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
     @Unroll
     def "validates gradle plugin and ccud plugin version"() {
         expect:
-        with(GlobalConfig.get().doCheckGradlePluginVersion(version)) {
+        with(InjectionConfig.get().doCheckGradlePluginVersion(version)) {
             kind == expectedKind
             message == expectedMssage
         }
-        with(GlobalConfig.get().doCheckCcudPluginVersion(version)) {
+        with(InjectionConfig.get().doCheckCcudPluginVersion(version)) {
             kind == expectedKind
             message == expectedMssage
         }
@@ -123,7 +122,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         def form = page.getFormByName("config")
 
         when:
-        form.getInputByName("_.injectionEnabled").click()
+        form.getInputByName("_.enabled").click()
         form.getInputByName("_.server").setValueAttribute("https://localhost")
         form.getInputByName("_.allowUntrusted").click()
         form.getInputByName("_.accessKey").setValueAttribute("ACCESS_KEY")
@@ -151,7 +150,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         def files = j.jenkins.root.listFiles(injectionConfigXmlFilter)
         files.length == 1
         with(fromXml(files.first().text)) {
-            injectionEnabled
+            enabled
             server == "https://localhost"
             allowUntrusted
             accessKey.plainText == "ACCESS_KEY"
@@ -177,7 +176,7 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         def form = page.getFormByName("config")
 
         when:
-        form.getInputByName("_.injectionEnabled").click()
+        form.getInputByName("_.enabled").click()
         form.getInputByName("_.server").setValueAttribute("https://localhost")
         form.getInputByName("_.allowUntrusted").click()
         form.getInputByName("_.accessKey").setValueAttribute(accessKey)
@@ -185,8 +184,8 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         j.submit(form)
 
         then:
-        def config = GlobalConfig.get()
-        config.injectionEnabled
+        def config = InjectionConfig.get()
+        config.enabled
         config.server == "https://localhost"
         config.accessKey == null
 
@@ -194,8 +193,8 @@ class InjectionConfigTest extends BaseGradleInjectionIntegrationTest {
         accessKey << ["", "   "]
     }
 
-    private static GlobalConfig fromXml(String xml) {
-        return (GlobalConfig) new XStream2().fromXML(xml)
+    private static InjectionConfig fromXml(String xml) {
+        return (InjectionConfig) new XStream2().fromXML(xml)
     }
 
     private static HtmlButton getAddButton(HtmlForm form, String label) {
